+++ vck-openid-ktor/src/commonMain/kotlin/at/asitplus/wallet/lib/ktor/openid/OAuth2KtorClient.kt
@@ -22,12 +22,15 @@ import at.asitplus.signum.indispensable.josef.JwsAlgorithm
 import at.asitplus.signum.indispensable.josef.JwsCompactTyped
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
+import at.asitplus.wallet.lib.agent.KeyMaterial
 import at.asitplus.wallet.lib.agent.RandomSource
 import at.asitplus.wallet.lib.jws.JwsHeaderCertOrJwk
+import at.asitplus.wallet.lib.jws.JwsHeaderNone
 import at.asitplus.wallet.lib.jws.SignJwt
 import at.asitplus.wallet.lib.jws.SignJwtFun
 import at.asitplus.wallet.lib.oauth2.OAuth2Client
 import at.asitplus.wallet.lib.oauth2.OAuth2Client.AuthorizationForToken
+import at.asitplus.wallet.lib.oidvci.BuildClientAttestationPoPJwt
 import at.asitplus.wallet.lib.oidvci.BuildDPoPHeader
 import at.asitplus.wallet.lib.oidvci.OAuth2Error
 import at.asitplus.wallet.lib.oidvci.OAuth2Exception.InvalidToken
@@ -69,8 +72,19 @@ class OAuth2KtorClient(
     cookiesStorage: CookiesStorage? = null,
     /** Additional configuration for building the HTTP client, e.g. callers may enable logging. */
     httpClientConfig: (HttpClientConfig<*>.() -> Unit)? = null,
-    /** Used to calculate DPoP, i.e. the key the access token and refresh token gets bound to. */
-    private val signDpop: SignJwtFun<JsonWebToken> = SignJwt(EphemeralKeyWithoutCert(), JwsHeaderCertOrJwk()),
+    /**
+     * Used to prove possession of the key material for the instance attestation.
+     * Also used for the DPoP signing function. (ts3-wallet-unit-attestation 1.5.1)
+     */
+    val keyMaterial: KeyMaterial = EphemeralKeyWithoutCert(),
+
+    /**
+     * DPoP sign function
+     * Uses instance attestation key material (ts3-wallet-unit-attestation 1.5.1)
+     **/
+    @Deprecated("Gets removed from the constructor in near future. Customization unnecessary because of ts3-wallet-unit-attestation 1.5.1")
+    private val signDpop: SignJwtFun<JsonWebToken> = SignJwt(keyMaterial = keyMaterial, JwsHeaderCertOrJwk()),
+
     /**
      * Implements OAuth2 protocol, `redirectUrl` needs to be registered by the OS for this application, so redirection
      * back from browser works
@@ -85,23 +99,13 @@ class OAuth2KtorClient(
 
     /** Returns a new instance attestation to validate the app against an authorization server. */
     val loadInstanceAttestation: (suspend (LoadInstanceAttestationInput) -> KmmResult<JwsCompactTyped<JsonWebToken>>)? = null,
-    /** Returns a proof of possession for an instance attestation */
-    val loadInstanceAttestationPop: (suspend (LoadInstanceAttestationPopInput) -> KmmResult<JwsCompactTyped<JsonWebToken>>)? = null,
-    @Deprecated("Use `loadInstanceAttestation` with `LoadInstanceAttestationInput`")
-    val loadClientAttestationJwt: (suspend () -> KmmResult<JwsCompactTyped<JsonWebToken>>)? = null,
-    @Deprecated("Use `loadInstanceAttestationPop` with `LoadInstanceAttestationPopInput`")
-    val signClientAttestationPop: (suspend () -> KmmResult<JwsCompactTyped<JsonWebToken>>)? = null,
-) {
+
+    ) {
     data class LoadInstanceAttestationInput(
         val authorizationServer: String,
         val preferredClientStatusPeriod: Duration?,
     )
 
-    data class LoadInstanceAttestationPopInput(
-        val authorizationServer: String,
-        val resourceUrl: String,
-        val httpMethod: HttpMethod,
-    )
 
     /**
      * Stores the latest DPoP nonce per origin. RFC 9449 requires using only the most recent nonce
@@ -498,10 +502,9 @@ class OAuth2KtorClient(
 
     /**
      * Sets the appropriate headers when accessing a token endpoint:
-     * - loads client attestation when [loadInstanceAttestation] and [loadInstanceAttestationPop] is set
+     * - loads client attestation when [loadInstanceAttestation] is set
      * - sends a DPoP proof when [useDpop] is set
      */
-    @Suppress("DEPRECATION")
     suspend fun applyAuthnForToken(
         resourceUrl: String,
         httpMethod: HttpMethod,
@@ -509,25 +512,29 @@ class OAuth2KtorClient(
         authorizationServer: String,
         preferredClientStatusPeriod: Duration?,
     ): HttpRequestBuilder.() -> Unit {
-        val clientAttJwt = if (loadInstanceAttestation != null || loadClientAttestationJwt != null) {
-            loadInstanceAttestation?.invoke(
-                LoadInstanceAttestationInput(
-                    authorizationServer = authorizationServer,
-                    preferredClientStatusPeriod = preferredClientStatusPeriod,
-                )
-            )?.getOrNull()?.jws
-                ?: loadClientAttestationJwt?.invoke()?.getOrNull()?.jws
-        } else null
-        val clientAttPop = if (loadInstanceAttestationPop != null || signClientAttestationPop != null) {
-            loadInstanceAttestationPop?.invoke(
-                LoadInstanceAttestationPopInput(
-                    authorizationServer = authorizationServer,
-                    resourceUrl = resourceUrl,
-                    httpMethod = httpMethod,
-                )
-            )?.getOrNull()?.jws
-                ?: signClientAttestationPop?.invoke()?.getOrNull()?.jws
-        } else null
+        val (clientAttJwt, clientAttPop) = when (loadInstanceAttestation != null) {
+            true -> {
+                loadInstanceAttestation.invoke(
+                    LoadInstanceAttestationInput(
+                        authorizationServer = authorizationServer,
+                        preferredClientStatusPeriod = preferredClientStatusPeriod,
+                    )
+                ).getOrElse { throw Exception("Unable to load instance attestation $it") }
+                    .jws to catching {
+                    BuildClientAttestationPoPJwt.invoke(
+                        signJwt = SignJwt(keyMaterial, JwsHeaderNone()),
+                        clientId = oAuth2Client.clientId,
+                        audience = authorizationServer,
+                        nonce = null // TODO: Add nonce after backend implementation is ready
+                    )
+                }.getOrElse { throw Exception("Unable to build instance attestation pop jwt $it") }
+                    .jws
+            }
+
+            else -> {
+                null to null
+            }
+        }
 
         val dpopHeader = useDpop.takeIf { it }?.let {
             BuildDPoPHeader(