+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientExternalAuthorizationServerTest.kt
@@ -69,10 +69,8 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
 
     data class Context(
         val credentialKeyMaterial: KeyMaterial,
-        val walletDpopKeyMaterial: KeyMaterial,
         val walletClientAuthKeyMaterial: KeyMaterial,
         val mockEngine: MockEngine,
-        val issuerDpopKeyMaterial: KeyMaterial,
         val issuerPublicContext: String,
         val issuerClientAuthKeyMaterial: KeyMaterial,
         val credentialIssuer: CredentialIssuer,
@@ -86,9 +84,7 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
         attributes: Map<String, String>,
     ): Context {
         val credentialKeyMaterial = EphemeralKeyWithoutCert()
-        val walletDpopKeyMaterial = EphemeralKeyWithoutCert()
         val walletClientAuthKeyMaterial = EphemeralKeyWithoutCert()
-        val issuerDpopKeyMaterial = EphemeralKeyWithoutCert()
         val issuerClientAuthKeyMaterial = EphemeralKeyWithoutCert()
         val credentialDataProvider = CredentialDataProviderFun {
             catching {
@@ -250,17 +246,7 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
                             )
                         }
                     },
-                    loadInstanceAttestationPop = {
-                        catching {
-                            BuildClientAttestationPoPJwt(
-                                SignJwt(issuerClientAuthKeyMaterial, JwsHeaderNone()),
-                                clientId = issuerPublicContext,
-                                audience = it.authorizationServer,
-                                lifetime = 10.minutes,
-                            )
-                        }
-                    },
-                    signDpop = SignJwt(issuerDpopKeyMaterial, JwsHeaderCertOrJwk()),
+                    keyMaterial = issuerClientAuthKeyMaterial,
                     oAuth2Client = OAuth2Client(clientId = issuerPublicContext),
                     randomSource = RandomSource.Default,
                 ),
@@ -274,10 +260,8 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
         )
         return Context(
             credentialKeyMaterial = credentialKeyMaterial,
-            walletDpopKeyMaterial = walletDpopKeyMaterial,
             walletClientAuthKeyMaterial = walletClientAuthKeyMaterial,
             mockEngine = mockEngine,
-            issuerDpopKeyMaterial = issuerDpopKeyMaterial,
             issuerPublicContext = issuerPublicContext,
             issuerClientAuthKeyMaterial = issuerClientAuthKeyMaterial,
             credentialIssuer = credentialIssuer,
@@ -300,17 +284,7 @@ val OpenId4VciClientExternalAuthorizationServerTest by testSuite {
                             )
                         }
                     },
-                    loadInstanceAttestationPop = {
-                        catching {
-                            BuildClientAttestationPoPJwt(
-                                SignJwt(walletClientAuthKeyMaterial, JwsHeaderNone()),
-                                clientId = walletClientId,
-                                audience = it.authorizationServer,
-                                lifetime = 10.minutes,
-                            )
-                        }
-                    },
-                    signDpop = SignJwt(walletDpopKeyMaterial, JwsHeaderCertOrJwk()),
+                    keyMaterial = walletClientAuthKeyMaterial,
                     oAuth2Client = OAuth2Client(clientId = walletClientId),
                     randomSource = RandomSource.Default,
                 )